changing root & paths
